+++ packages/common/src/Config.ts
@@ -16,6 +16,7 @@ export interface CCFConfig {
     ATHENA_DB_TABLE?: string
     ATHENA_QUERY_RESULT_LOCATION?: string
     ATHENA_REGION?: string
+    IS_AWS_GLOBAL?: boolean
     NAME?: string
     RECOMMENDATIONS_SERVICE?: AWS_RECOMMENDATIONS_SERVICES
     COMPUTE_OPTIMIZER_BUCKET?: string
@@ -55,6 +56,8 @@ export interface CCFConfig {
       clientSecret?: string
       tenantId?: string
     }
+    RESOURCE_TAG_NAMES?: string[]
+    CONSUMPTION_CHUNKS_DAYS?: number
   }
   LOGGING_MODE?: string
   CACHE_MODE?: string
@@ -95,6 +98,12 @@ export type QUERY_DATE_TYPES = {
   [key in GroupBy]: string
 }
 
+// this check allows for aws auth to determine how to correctly partition by region when credentializing
+const checkAthenaRegionISAWSGlobal = (athena_region: string): boolean => {
+  const AWS_CN_REGIONS = ['cn-north-1', 'cn-northwest-1']
+  return !AWS_CN_REGIONS.includes(athena_region)
+}
+
 const getAWSAccounts = () => {
   return process.env.AWS_ACCOUNTS ? process.env.AWS_ACCOUNTS : '[]'
 }
@@ -102,6 +111,12 @@ const getAWSAccounts = () => {
 const getAWSResourceTagNames = () => {
   return process.env.AWS_RESOURCE_TAG_NAMES
     ? process.env.AWS_RESOURCE_TAG_NAMES
+    : '[]'
+}
+
+const getAzureResourceTagNames = () => {
+  return process.env.AZURE_RESOURCE_TAG_NAMES
+    ? process.env.AZURE_RESOURCE_TAG_NAMES
     : '[]'
 }
 
@@ -133,6 +148,7 @@ const getConfig = (): CCFConfig => ({
     ATHENA_QUERY_RESULT_LOCATION:
       getEnvVar('AWS_ATHENA_QUERY_RESULT_LOCATION') || '',
     ATHENA_REGION: getEnvVar('AWS_ATHENA_REGION') || '',
+    IS_AWS_GLOBAL: checkAthenaRegionISAWSGlobal(getEnvVar('AWS_ATHENA_REGION')),
     accounts: JSON.parse(getAWSAccounts()) || [],
     authentication: {
       mode: getEnvVar('AWS_AUTH_MODE') || 'default',
@@ -188,6 +204,10 @@ const getConfig = (): CCFConfig => ({
         key: 'rds',
         name: 'RDS',
       },
+      {
+        key: 'lambda',
+        name: 'Lambda',
+      },
     ],
   },
   GCP: {
@@ -238,6 +258,10 @@ const getConfig = (): CCFConfig => ({
       clientSecret: getEnvVar('AZURE_CLIENT_SECRET') || '',
       tenantId: getEnvVar('AZURE_TENANT_ID') || '',
     },
+    RESOURCE_TAG_NAMES: JSON.parse(getAzureResourceTagNames()),
+    CONSUMPTION_CHUNKS_DAYS: parseInt(
+      getEnvVar('AZURE_CONSUMPTION_CHUNKS_DAYS') || '0',
+    ),
   },
   LOGGING_MODE: process.env.LOGGING_MODE || '',
   CACHE_MODE: getEnvVar('CACHE_MODE') || '',